storage/image-requirements.txt
numpy==1.19.5
paho_mqtt==1.6.1
Pillow==8.4.0

